Move the manager into a separate file.
