+++ emmaa/db/schema.py
@@ -1,9 +1,11 @@
+__all__ = ['User', 'Query', 'UserQuery', 'Result']
+
 import logging
 
-from sqlalchemy.ext.declarative import declarative_base, DeclarativeMeta
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Column, Integer, String, UniqueConstraint, ForeignKey, \
-    create_engine, inspect, LargeBinary, Boolean, DateTime, func, BigInteger
-from sqlalchemy.orm import relationship, sessionmaker
+    Boolean, DateTime, func, BigInteger
+from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import JSONB
 
 
@@ -54,6 +56,7 @@ class UserQuery(Base, EmmaaTable):
     query_hash = Column(BigInteger, ForeignKey('query.hash'), nullable=False)
     query = relationship(Query)
     date = Column(DateTime, default=func.now())
+    subscription = Column(Boolean, nullable=False)
 
 
 class Result(Base, EmmaaTable):
@@ -62,45 +65,3 @@ class Result(Base, EmmaaTable):
     query_hash = Column(BigInteger, ForeignKey('query.hash'), nullable=False)
     date = Column(DateTime, default=func.now())
     json = Column(JSONB, nullable=False)
-
-
-class EmmaaDatabaseError(Exception):
-    pass
-
-
-class EmmaaDatabaseManager(object):
-    table_order = ['user', 'query', 'user_query', 'result']
-
-    def __init__(self, host):
-        self.host = host
-        self.engine = create_engine(host)
-        self.tables = {tbl.__tablename__: tbl
-                       for tbl in EmmaaTable.__subclasses__()}
-        self.session = None
-        return
-
-    def grab_session(self):
-        if self.session is None or not self.session.is_active:
-            logger.debug(f"Grabbing a session to {self.host}...")
-            DBSession = sessionmaker(bind=self.engine)
-            logger.debug("Session grabbed.")
-            self.session = DBSession()
-            if self.session is None:
-                raise EmmaaDatabaseError("Could not acquire session.")
-
-    def create_tables(self, tables=None):
-        if tables is None:
-            tables = set(self.tables.keys())
-        else:
-            tables = set(tables)
-
-        for tbl_name in self.table_order:
-            if tbl_name in tables:
-                logger.info(f"Creating {tbl_name} table")
-                if not self.tables[tbl_name].__table__.exists(self.engine):
-                    self.tables[tbl_name].__table__.create(bind=self.engine)
-                    logger.debug("Table created!")
-                else:
-                    logger.warning(f"Table {tbl_name} already exists! "
-                                   f"No action taken.")
-        return
